Refactor all etl tasks to io filepaths
Now every task in residential etl takes a filepath input,
performs some operation & saves the result in an output
filepath.  This means that each task doesn't depend on
global state and so can be run individually if required -
run flow to generate data and can then play with the tasks
independent of flow context...

Also mocks out DATA_DIR for FTEST_DATA_DIR instead of passing
it as a flow parameter (to residential etl) as this produces a
cleaner etl visualization.  Previously the external_dir prefect
parameter was split in the viz into a number of Div bubbles rendering
the etl viz into a dependency hell mess.  Now each filepath input
appears in the viz as independent to all others and so the dependency
mess problem is eliminated...